Add field errors to registration form template
Before, it was possible to enter invalid data into the registration
form, e.g. duplicate username, insecure password, etc., without being
notified about the violation. The form would simply reload and no user
is created and no email is sent.

Without being notified about the registration form issue, one might
assume that something in the setup is not correct and waste hours
trying to debug something that is not broken. At least, that is what
happened to me 😬

This commit adds the form field errors to the registration form
template.