Fix order of API view functions
Before, the routes/views with parameters in the URL where defined first.
E.g. `@api.route("/api/movie/genre/{genre}")` was defined before
`@api.route("/api/movie/genre/all")`.

When trying to hit the URL for `/api/movie/genre/all` the response would
be empty. This is because the this would not trigger the desired view,
but the parameterized view would be passes `genre = all`. Since there
are no movies with the genre "all" the response would be empty.

To solve this issue, the ordering of the routes/views had to be
adjusted. Now the most specific views are defined first, and then the
more flexible views are defined.